[TM-3193] Align Review Details tab with form design
