+++ src/components/extensive/PageElements/PageContent/components/PhotosAndVideosSection.tsx
@@ -8,12 +8,13 @@ import {
 } from "@/components/extensive/WizardForm/FormSummaryRow/parseFilesFromHtml";
 import { PlayCircleIcon } from "@/redesignComponents/foundations/Icons";
 import SimpleDivider from "@/redesignComponents/miscellaneous/Dividers/SimpleDivider";
+import { EntityName, SingularEntityName } from "@/types/common";
 
 import MediaSection from "./MediaSection";
 
 type PhotosAndVideosSectionProps = {
   value: string;
-  entityName?: "projects" | "sites";
+  entityName?: EntityName | SingularEntityName;
   entityUUID?: string;
 };
