+++ src/components/extensive/PageElements/PageContent/components/sharedDetails.tsx
@@ -1,9 +1,9 @@
-import { Box, Flex, TableCell, TableRow, Text } from "@chakra-ui/react";
+import { Flex, Text } from "@chakra-ui/react";
 import { useT } from "@transifex/react";
 import classNames from "classnames";
 import { Dictionary } from "lodash";
 import { useRouter } from "next/router";
-import { FC, Fragment, useMemo } from "react";
+import { FC, Fragment } from "react";
 
 import { formatEntryValue } from "@/admin/apiProvider/utils/entryFormat";
 import { PLANTING_STATUS_MAP } from "@/components/elements/Status/constants/statusMap";
@@ -14,26 +14,15 @@ import { FormStepWithValidation } from "@/components/extensive/WizardForm/useFor
 import { ProjectFullDto, SiteFullDto } from "@/generated/v3/entityService/entityServiceSchemas";
 import { isEntityAwaitingApproval, v3EntityName } from "@/helpers/entity";
 import { useGetEditEntityHandler } from "@/hooks/entity/useGetEditEntityHandler";
-import {
-  COUNT_TABLE_SPECIES_PER_PAGE_MIN,
-  getPlantingStatus,
-  NO_COUNT_TABLE_SPECIES_PER_PAGE,
-  NO_COUNT_TABLE_SPECIES_PER_ROW,
-  noCountTableColumns
-} from "@/pages/project/[uuid]/tabs/constants/Detail.constants";
+import { getPlantingStatus } from "@/pages/project/[uuid]/tabs/constants/Detail.constants";
 import Button from "@/redesignComponents/actions/Buttons/Button/Button";
 import { ProgressTag } from "@/redesignComponents/actions/Tags/ProgressTag/ProgressTag";
 import Accordion from "@/redesignComponents/containers/Accordion/Accordion";
 import AccordionHeader from "@/redesignComponents/containers/Accordion/AccordionHeader";
-import Table from "@/redesignComponents/dataDisplay/Table/Table";
-import {
-  FULL_WIDTH_TABLE_HEADER_STYLES,
-  NO_HEADER_TABLE_WRAPPER_STYLES
-} from "@/redesignComponents/dataDisplay/Table/tableStyles";
 import { ArrowForward, EditIcon } from "@/redesignComponents/foundations/Icons";
 
 import { getFieldsRequiringAttentionCount, plantsToNoCountRows } from "../utils/detailUtils";
-import SpecialEntryRenderer, { SPECIAL_ENTRY_TITLES } from "./SpecialEntryRenderer";
+import SpecialEntryRenderer, { PlantTableEntryRenderer, SPECIAL_ENTRY_TITLES } from "./SpecialEntryRenderer";
 
 export { getFieldsRequiringAttentionCount, plantsToNoCountRows };
 
@@ -56,12 +45,9 @@ export type SharedDetailsProps = {
 
 type EntryValueRendererProps = {
   entry: FormEntry;
-  noGoalTableColumns: { key: string; label: string }[];
-  entityName: "projects" | "sites";
-  entityUUID: string;
 };
 
-const EntryValueRenderer = ({ entry, noGoalTableColumns, entityName, entityUUID }: EntryValueRendererProps) => {
+const EntryValueRenderer = ({ entry }: EntryValueRendererProps) => {
   const rawValue = entry.value ?? "-";
 
   if (typeof rawValue === "string" || typeof rawValue === "number") {
@@ -70,61 +56,8 @@ const EntryValueRenderer = ({ entry, noGoalTableColumns, entityName, entityUUID 
     );
   }
 
-  if (rawValue.props.tableType === "noCount") {
-    const noCountTableRowCount = rawValue.props.plants.length / NO_COUNT_TABLE_SPECIES_PER_ROW;
-    const dataPlants = plantsToNoCountRows(rawValue.props.plants);
-
-    return (
-      <Table
-        data={dataPlants}
-        columns={noCountTableColumns}
-        css={NO_HEADER_TABLE_WRAPPER_STYLES}
-        variant="full-width"
-        totalItems={noCountTableRowCount}
-        showItemCount={false}
-        pageSize={NO_COUNT_TABLE_SPECIES_PER_PAGE}
-        showPagination={NO_COUNT_TABLE_SPECIES_PER_PAGE < noCountTableRowCount}
-        className={classNames("mt-[2px]", dataPlants.length <= NO_COUNT_TABLE_SPECIES_PER_PAGE && "mb-3")}
-        renderRow={rowData => {
-          const row = rowData as Record<number, string> & { id: number };
-          return (
-            <TableRow>
-              {noCountTableColumns.map((col, idx) => (
-                <TableCell key={col.key + idx} className={idx === 0 ? undefined : "px-0! py-4"}>
-                  {row[idx + 1] !== undefined && row[idx + 1] !== "" && (
-                    <Box
-                      className={classNames(
-                        "border-b border-theme-neutral-300 py-4",
-                        idx === noCountTableColumns.length - 1 ? "" : "mr-8"
-                      )}
-                    >
-                      {row[idx + 1]}
-                    </Box>
-                  )}
-                </TableCell>
-              ))}
-            </TableRow>
-          );
-        }}
-      />
-    );
-  }
-
-  if (rawValue.props.tableType === "noGoal") {
-    return (
-      <Table
-        data={rawValue.props.plants}
-        columns={noGoalTableColumns}
-        variant="full-width"
-        css={FULL_WIDTH_TABLE_HEADER_STYLES}
-        totalItems={rawValue.props.plants.length}
-        showItemCount={false}
-        className={classNames(
-          "mt-[2px] !w-[725px]",
-          rawValue.props.plants.length <= COUNT_TABLE_SPECIES_PER_PAGE_MIN && "mb-3"
-        )}
-      />
-    );
+  if (rawValue.props.tableType === "noCount" || rawValue.props.tableType === "noGoal") {
+    return <PlantTableEntryRenderer rawValue={rawValue} />;
   }
 
   return (
@@ -164,14 +97,6 @@ const SharedDetails: FC<SharedDetailsProps> = ({
     entityStatus: entityStatus ?? "started",
     updateRequestStatus: updateRequestStatus ?? "no-update"
   });
-
-  const noGoalTableColumns = useMemo(
-    () => [
-      { key: "name", label: t("Species Name") },
-      { key: "amount", label: t("Number of Trees Expected") }
-    ],
-    [t]
-  );
 
   return entries.length === 0 ? null : (
     <Accordion
@@ -235,7 +160,7 @@ const SharedDetails: FC<SharedDetailsProps> = ({
             </Flex>
           );
 
-          if (SPECIAL_ENTRY_TITLES.has(entry.title ?? "")) {
+          if (SPECIAL_ENTRY_TITLES.has(entry.title ?? "") || entry.inputType === "treeSpecies") {
             return (
               <Fragment key={`${step.id}-${entry.title}-${index}`}>
                 <SpecialEntryRenderer entry={entry} entityName={entityName} entityUUID={entityUUID} />
@@ -258,12 +183,7 @@ const SharedDetails: FC<SharedDetailsProps> = ({
                 <div
                   className={classNames("my-2 h-px w-full bg-theme-neutral-300", !isAdditionalInformation && "hidden")}
                 />
-                <EntryValueRenderer
-                  entry={entry}
-                  noGoalTableColumns={noGoalTableColumns}
-                  entityName={entityName}
-                  entityUUID={entityUUID}
-                />
+                <EntryValueRenderer entry={entry} />
               </Flex>
               {projectStageSection}
             </Fragment>
